Move "TcpLogSink" next to "TcpLogger" & co and same namespace
